manifesto: Align string descriptor's end to four byte boundary
We believe that we are doing this, but we weren't. We are adding N
characters with value "0x00" at the end of the string, where N is:
'strlen % 4'. But what we need to do is: '4 - strlen % 4'

Also we need to take size of the descriptor as well into consideration
here, which was missing.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>