manifesto: parse ID from section headers
A few descriptors in the .mnfs format currently look like this:

    [foo-descriptor "bar"]
    foo-field = 0xdeadbeef
    id = <ID>

The "bar" in the foo-descriptor section header needs to be unique in the
file, but is ignored by the parser. In practice, people are using "<ID>"
in there, though.

To reduce redundancy and potential for error, parse the ID from the
field that follows foo-descriptor. You can use either this syntax:

    [foo-descriptor "id"]
    foo-field = 0xdeadbeef

Or this one:

    [foo-descriptor id]
    foo-field = 0xdeadbeef

In addition, emit a warning when old-style "id" items are found, so
people know to fix up their old manifests.

All manifests in the examples/ directory have identical binary output
before and after this patch.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>