SW-3535: new version of Manifesto
The manifesto tool is currently reduced to its minimum: it generates a binary
version of a manifest file.

It performs very little checking and for example, doesn't make sure that the
class and protocol numbers (for bundle and cport descriptors) are valid, that
cport descriptors refer to existing bundles, etc.

Errors in the manifest file are unfortunately difficult to track once the file
is compiled in a firmware image and a few partners had issues because of that.

This new version is a rework that should resolve most of these issues.

1/ the command line options are enhanced (inspired by the 'dtc' tool):

usage: manifesto [-h] [-I {mnfs,mnfb}] [-o OUT] [-O {mnfs,mnfb}] [-s] infile

positional arguments:
  infile                specify an input manifest file

optional arguments:
  -h, --help            show this help message and exit
  -I {mnfs,mnfb}, --in-format {mnfs,mnfb}
                        input formats (default is mnfs)
  -o OUT, --out OUT     output file (default is stdout)
  -O {mnfs,mnfb}, --out-format {mnfs,mnfb}
                        output formats (default is mnfb)
  -s, --silent          silence the (potentially relevant) warnings

(Note that the mnfb input format is currently unimplemented.)

2/ performs error checking. Among others, it detects:
- invalid string ids == 0
- multiple references to one string
- descriptors with same id (configparser doesn't always get it right because of
  possible spacing between the descriptor type and the id)
- invalid id references from interface descriptor to strings
- invalid bundle references from cport descriptors to bundle descriptors
- invalid class for bundle 0
- invalid protocol for cport 0
- out of range values for id/class/protocol/etc.
- many more...

3/ gives warnings:
- "reserved" class or protocol
- incompatible protocol with given bundle class
- missing control cport/bundle
- unused descriptor
- non-incremental id for bundles

Testing done:
- generated all the mnfb of all the examples and compared the binary output
  (md5sum) with what the old tool generated
- generated all the mnfb of the all the FDK modules and compared the binary
  output (md5sum) with what the old tool generated

Change-Id: Ib35126917be9cc6c9158292c681cd8570fdb44ee
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>