manifesto : new descriptors for mikrobus
mikroBUS is an add-on board socket standard by
MikroElektronika which provides high expandability
with the lowest number of pins used.

The mikrobus linux kernel driver makes use of an
extended version of the greybus manifest to pass
the device driver specific information, the manifest
binary is expected to be stored on an on-board EEPROM
from where the mikroBUS driver fetches the contents and
enumerates the drivers.